+++ changelog.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 * Changed: Added in more formal documentation and reasoning behind Lazy Loading classes in our Service Providers.
 * Changed: Updated default PHP version to 7.2
 * Changed: Added in default Panels Caching
+* Changed: Added in Whoops library
 
 ## [2019.03]